login

+++ QuanLyDonVi/QuanLyDonVi/GUI/Uc_QLTaiKhoan.cs
@@ -16,12 +16,14 @@ namespace QuanLyDonVi.GUI
     public partial class Uc_QLTaiKhoan : DevExpress.XtraEditors.XtraUserControl
     {
         BindingSource list = new BindingSource();
-        public Uc_QLTaiKhoan()
+        Account acc = new Account();
+        public Uc_QLTaiKhoan(Account _acc)
         {
             InitializeComponent();
             LoadAccount();
             dgrAccount.DataSource = list;
             Binding();
+            acc = _acc;
         }
 
         private void LoadAccount()
@@ -32,6 +34,7 @@ namespace QuanLyDonVi.GUI
         private void Binding()
         {
             txbUsername.DataBindings.Add("Text", dgrAccount.DataSource, "Username", true, DataSourceUpdateMode.Never);
+            cboQuyen.DataBindings.Add("Text", dgrAccount.DataSource, "Type", true, DataSourceUpdateMode.Never);
         }
 
         private void btn_Them_Click(object sender, EventArgs e)
@@ -46,29 +49,72 @@ namespace QuanLyDonVi.GUI
             }
             else
             {
-                if(txbPass.Text == txbRePass.Text)
+                if(acc.Type == "Root")
                 {
-                    Account item = new Account();
-                    item.Username = txbUsername.Text;
-                    item.Password = txbPass.Text; 
-                    if (new AccountDAO().Add(item))
+                    if (txbPass.Text == txbRePass.Text)
                     {
-                        LoadAccount();
-                        MessageBox.Show("Thêm thành công");
+                        Account item = new Account();
+                        item.Username = txbUsername.Text;
+                        item.Password = txbPass.Text;
+                        item.Type = cboQuyen.Text;
+                        if (new AccountDAO().Add(item))
+                        {
+                            LoadAccount();
+                            MessageBox.Show("Thêm thành công");
+                        }
+                        else
+                        {
+                            MessageBox.Show("Không thành công");
+                        }
+
+                        pnlPass.Visible = false;
+                        txbUsername.ReadOnly = true;
+                        btn_Them.Text = "Thêm tài khoản";
+                        btnXoa.Text = "Xóa tài khoản";
                     }
                     else
                     {
-                        MessageBox.Show("Không thành công");
+                        MessageBox.Show("Xác nhận mật khẩu không chính xác");
                     }
+                }
+                else if(acc.Type == "Admin")
+                {
+                    if (cboQuyen.Text == "Admin" || cboQuyen.Text == "User")
+                    {
+                        if (txbPass.Text == txbRePass.Text)
+                        {
+                            Account item = new Account();
+                            item.Username = txbUsername.Text;
+                            item.Password = txbPass.Text;
+                            item.Type = cboQuyen.Text;
+                            if (new AccountDAO().Add(item))
+                            {
+                                LoadAccount();
+                                MessageBox.Show("Thêm thành công");
+                            }
+                            else
+                            {
+                                MessageBox.Show("Không thành công");
+                            }
 
-                    pnlPass.Visible = false;
-                    txbUsername.ReadOnly = true;
-                    btn_Them.Text = "Thêm tài khoản";
-                    btnXoa.Text = "Xóa tài khoản";
+                            pnlPass.Visible = false;
+                            txbUsername.ReadOnly = true;
+                            btn_Them.Text = "Thêm tài khoản";
+                            btnXoa.Text = "Xóa tài khoản";
+                        }
+                        else
+                        {
+                            MessageBox.Show("Xác nhận mật khẩu không chính xác");
+                        }
+                    }
+                    else
+                    {
+                        MessageBox.Show("Bạn không có quyền thực hiện tác vụ này");
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("Xác nhận mật khẩu không chính xác");
+                    MessageBox.Show("Bạn không có quyền thực hiện tác vụ này");
                 }
             }
         }
@@ -84,12 +130,37 @@ namespace QuanLyDonVi.GUI
             }
             else
             {
-                if (MessageBox.Show("Bạn có muốn xóa không", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                if(acc.Type == "Root")
                 {
-                    if(new AccountDAO().Remove(txbUsername.Text))
+                    if (MessageBox.Show("Bạn có muốn xóa không", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                     {
-                        LoadAccount();
+                        if (new AccountDAO().Remove(txbUsername.Text))
+                        {
+                            LoadAccount();
+                        }
                     }
+                }
+                else if(acc.Type == "Admin")
+                {
+                    string cur_type = new AccountDAO().GetTypeByU(acc.Username);
+                    if(cur_type == "Admin" || cur_type == "User")
+                    {
+                        if (MessageBox.Show("Bạn có muốn xóa không", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                        {
+                            if (new AccountDAO().Remove(txbUsername.Text))
+                            {
+                                LoadAccount();
+                            }
+                        }
+                    }
+                    else
+                    {
+                        MessageBox.Show("Bạn không có quyền thực hiện tác vụ này");
+                    }
+                }
+                else
+                {
+                    MessageBox.Show("Bạn không có quyền thực hiện tác vụ này");
                 }
             }
         }
